+++ src/main/java/subway/application/LineService.java
@@ -3,12 +3,12 @@ package subway.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import subway.controller.dto.LineRequest;
+import subway.controller.dto.LineResponse;
 import subway.dao.SectionDao;
 import subway.domain.Line;
 import subway.repository.LineRepository;
 import subway.repository.StationRepository;
-import subway.ui.dto.LineRequest;
-import subway.ui.dto.LineResponse;
 
 @Service
 public class LineService {
