+++ src/main/java/subway/controller/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.controller.dto;
 
 public class LineRequest {
