+++ src/main/java/subway/controller/dto/LineStationResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.controller.dto;
 
 import java.util.List;
