+++ src/main/java/subway/controller/dto/SectionDeleteRequest.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.controller.dto;
 
 public class SectionDeleteRequest {
