+++ src/main/java/subway/controller/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway.ui.dto;
+package subway.controller.dto;
 
 import subway.domain.Station;
