+++ src/main/java/subway/dao/LineDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.dao.dto.LineEntity;
+import subway.dao.entity.LineEntity;
 
 @Repository
 public class LineDao {
