+++ src/main/java/subway/dao/SectionDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.dao.dto.SectionEntity;
+import subway.dao.entity.SectionEntity;
 
 @Repository
 public class SectionDao {
