+++ src/main/java/subway/dao/StationDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
-import subway.dao.dto.StationEntity;
+import subway.dao.entity.StationEntity;
 
 @Component
 public class StationDao {
