+++ src/main/java/subway/dao/entity/SectionEntity.java
@@ -1,4 +1,4 @@
-package subway.dao.dto;
+package subway.dao.entity;
 
 public class SectionEntity {
