+++ src/main/java/subway/repository/LineRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import org.springframework.stereotype.Repository;
 import subway.dao.LineDao;
 import subway.dao.SectionDao;
-import subway.dao.dto.LineEntity;
-import subway.dao.dto.SectionEntity;
+import subway.dao.entity.LineEntity;
+import subway.dao.entity.SectionEntity;
 import subway.domain.Distance;
 import subway.domain.Line;
 import subway.domain.Section;
