+++ src/test/java/subway/integration/StationIntegrationTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
-import subway.ui.dto.StationResponse;
+import subway.controller.dto.StationResponse;
 
 @Sql("/truncate.sql")
 @DisplayName("지하철역 관련 기능")
